// src/store/types/PostsActionType.ts
export enum PostsActionType {
    FETCH_POSTS = 'POSTS_FETCH_POSTS',
    RESET_POSTS = 'POSTS_RESET_POSTS',
    START_CREATING_NEW_POST = 'POSTS_START_CREATING_NEW_POST',
    CANCEL_CREATING_NEW_POST = 'POSTS_CANCEL_CREATING_NEW_POST',
    SUBMIT_NEW_POST = 'POSTS_SUBMIT_NEW_POST',
    SET_POSTS_PAGE = 'POSTS_SET_POSTS_PAGE',
    SET_SEARCH_TITLE = 'POSTS_SET_SEARCH_TITLE',
}